requirements.txt
selenium==4.15.2
python-dotenv